rework willkommen based on template
