+++ willkommen/.cs50.yml
@@ -6,3 +6,5 @@ submit50:
   
 check50:
   files: *submit50_files
+  checks: checks/main.py
+
